requirements.txt

pymsql
mysql-connector-python